Move attributes related to configuration of an event or consumer to the Configurations namespace
