+++ src/Tingle.EventBus/Configuration/ConsumerNameAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tingle.EventBus
+namespace Tingle.EventBus.Configuration
 {
     /// <summary>
     /// Specify the ConsumerName used for this consumer type, overriding the generated one.
