+++ src/Tingle.EventBus/Configuration/EventNameAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tingle.EventBus
+namespace Tingle.EventBus.Configuration
 {
     /// <summary>
     /// Specify the EventName used for this event contract/type, overriding the generated one.
