+++ src/Tingle.EventBus/Configuration/EventTransportNameAttribute.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Tingle.EventBus
+namespace Tingle.EventBus.Configuration
 {
     /// <summary>
     /// Specify the name of the transport used for this event contract/type, overriding the default one.
