+++ tests/Tingle.EventBus.Tests/Configurator/TestConsumer3.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
+using Tingle.EventBus.Configuration;
 
 namespace Tingle.EventBus.Tests.Configurator
 {
