+++ tests/Tingle.EventBus.Tests/Configurator/TestEvent3.cs
@@ -1,4 +1,6 @@
-﻿namespace Tingle.EventBus.Tests.Configurator
+﻿using Tingle.EventBus.Configuration;
+
+namespace Tingle.EventBus.Tests.Configurator
 {
     [EntityKind(EntityKind.Broadcast)]
     [EventSerializer(typeof(FakeEventSerializer2))]
